refactor: split store.test.ts into focused test files
